voice_agent/pyproject.toml
[project]
name = "agentic-ai-voice-agent"
version = "0.1.0"
description = "Voice Agentic AI Platform - LiveKit Voice Agent Service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit>=0.17.0",
    "livekit-agents>=0.9.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-openai>=0.7.0",
    "livekit-plugins-silero>=0.6.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.3",
    "langgraph>=0.0.20",
    "qdrant-client>=1.7.0",
    "pypdf>=3.17.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "duckduckgo-search>=4.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
